% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Density of the triangular distribution}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{Numeric vector where the density is evaluated.}

\item{min}{Lower limit of the distribution.}

\item{max}{Upper limit of the distribution.}

\item{mode}{Mode of the distribution (between min and max).}
}
\value{
Numeric vector with density values.
}
\description{
Density of the triangular distribution
}
