% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Cumulative distribution function of the triangular distribution}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{Numeric vector of quantiles.}

\item{min}{Lower limit.}

\item{max}{Upper limit.}

\item{mode}{Mode of the distribution.}
}
\value{
Numeric vector with CDF values.
}
\description{
Cumulative distribution function of the triangular distribution
}
