% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Random generation from the triangular distribution}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{Number of observations.}

\item{min}{Lower limit.}

\item{max}{Upper limit.}

\item{mode}{Mode of the distribution.}
}
\value{
Numeric vector of length n.
}
\description{
Random generation from the triangular distribution
}
